requirements.txt
whisperx==3.1.1
pyannote.audio==3.1.1
pyannote.core==5.0.0
torch==2.4.0
torchaudio==2.4.0
ffmpeg-python==0.2.0
huggingface-hub==0.26.1
black==24.10.0
python-dotenv==1.0.1